b3 refactoring reviews.js
